Move annotator.plugin.filter -> annotator.ui.filter.standalone
