+++ test/spec/ui/filter_spec.js
@@ -402,3 +402,52 @@ describe('ui.filter.Filter', function () {
         });
     });
 });
+
+
+describe('ui.filter.standalone', function () {
+    var mockFilter = null,
+        plugin = null,
+        sandbox = null;
+
+    beforeEach(function () {
+        sandbox = sinon.sandbox.create();
+        mockFilter = {
+            updateHighlights: sandbox.stub(),
+            destroy: sandbox.stub()
+        };
+
+        sandbox.stub(filter, 'Filter').returns(mockFilter);
+
+        plugin = filter.standalone();
+    });
+
+    afterEach(function () {
+        sandbox.restore();
+    });
+
+    var hooks = [
+        'annotationsLoaded',
+        'annotationCreated',
+        'annotationUpdated',
+        'annotationDeleted'
+    ];
+
+    function testHook(h) {
+        return function () {
+            plugin[h]({text: 123});
+            sinon.assert.calledWith(mockFilter.updateHighlights);
+        };
+    }
+
+    for (var i = 0, len = hooks.length; i < len; i++) {
+        it(
+            "calls updateHighlights on the filter component " + hooks[i],
+            testHook(hooks[i])
+        );
+    }
+
+    it('destroys the filter component when destroyed', function () {
+        plugin.destroy();
+        sinon.assert.calledOnce(mockFilter.destroy);
+    });
+});
